move Fetch-Callbacks to own file
