+++ libGitWrap/CloneOperationPrivate.hpp
@@ -55,18 +55,8 @@ namespace Git
             git_remote_callbacks    mRemoteCallbacks;
 
             Result                  mResult;
-
-        private:
-            static int  credAccquire(git_cred** cred, const char* url, const char *username_from_url,
-                                      unsigned int allowed_types, void* payload );
-            static int  fetchProgress( const git_transfer_progress* stats, void* payload );
-            static int  remoteComplete( git_remote_completion_type type, void* payload );
-            static void remoteProgress( const char* str, int len, void* payload );
-            static int  remoteUpdateTips( const char* refname, const git_oid* a, const git_oid* b,
-                                          void* payload );
         };
 
     }
-
 
 }
